aarch64: Fix IMAGE_REL_ARM64_PAGEBASE_REL21 relocation and symbol reduction for relocations
Initially, the change was to fix IMAGE_REL_ARM64_PAGEBASE_REL21,
which was addressing only 1MB instead of 4GB. This change triggered
multiple changes in binutils and gcc to support the small code model.

Probably, rel->r_type = IMAGE_REL_ARM64_ABSOLUTE was done on purpose
to bypass changes in processed relocations in a later call to
_bfd_coff_generic_relocate_section. However, it was breaking the
relocation reduction logic, as the initial relocation type was no
longer known because PE relocations have been replaced with
IMAGE_REL_ARM64_ABSOLUTE. To avoid extra processing and introducing
an extra field in the relocation structure, rel->r_type =
IMAGE_REL_ARM64_ABSOLUTE has been replaced with rel->r_vaddr = -1,
and a check to skip relocation processing has been added to
_bfd_coff_generic_relocate_section.

In one of the relocation processing steps, there is logic to reduce
a symbol name to a section name where it is placed. In the COFF
format on AArch64, this should not happen for symbols used in
relocations that cannot be fully handled at the object compiling
time. This information will be used by the linker to make correct
relocation. Otherwise, it can lead to a silent relocation overflow
during the linking step because the linker assumes everything looks
good, which most likely results in a crash upon execution.

To resolve this issue, S_SHOULD_BE_REDUCED_TO_SECTION_NAME has been
added.

This patch supports the GCC patch series 3 and should be merged after
upstreaming that patch series.

bfd/ChangeLog:

	* coff-aarch64.c (coff_aarch64_rel21_reloc): Fix relocation
	(coff_pe_aarch64_relocate_section): Likewise.
	* cofflink.c (_bfd_coff_generic_relocate_section): Likewise.

gas/ChangeLog:

	* symbols.c (S_SHOULD_BE_REDUCED_TO_SECTION_NAME): New.
	(defined): New.
	* symbols.h (S_SHOULD_BE_REDUCED_TO_SECTION_NAME): New.
	* write.c (adjust_reloc_syms): Use
	S_SHOULD_BE_REDUCED_TO_SECTION_NAME.